Add BugsnagInternals.h
