+++ Bugsnag/Bugsnag.m
@@ -30,6 +30,7 @@
 #import "Bugsnag+Private.h"
 #import "BugsnagBreadcrumbs.h"
 #import "BugsnagClient+Private.h"
+#import "BugsnagInternals.h"
 #import "BugsnagLogger.h"
 
 static BugsnagClient *bsg_g_bugsnag_client = NULL;
@@ -66,17 +67,6 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
  */
 + (void)purge {
     bsg_g_bugsnag_client = nil;
-}
-
-+ (BugsnagConfiguration *)configuration {
-    if ([self bugsnagStarted]) {
-        return self.client.configuration;
-    }
-    return nil;
-}
-
-+ (BugsnagConfiguration *)instance {
-    return [self configuration];
 }
 
 + (BugsnagClient *)client {
@@ -124,18 +114,6 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
 + (void)notifyError:(NSError *)error block:(BugsnagOnErrorBlock)block {
     if ([self bugsnagStarted]) {
         [self.client notifyError:error block:block];
-    }
-}
-
-/**
- * Intended for use by other clients (React Native/Unity). Calling this method
- * directly from iOS is not supported.
- */
-+ (void)notifyInternal:(BugsnagEvent *_Nonnull)event
-                 block:(BugsnagOnErrorBlock)block {
-    if ([self bugsnagStarted]) {
-        [self.client notifyInternal:event
-                              block:block];
     }
 }
 
@@ -198,13 +176,6 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
         return [self.client resumeSession];
     } else {
         return false;
-    }
-}
-
-+ (void)addRuntimeVersionInfo:(NSString *)info
-                      withKey:(NSString *)key {
-    if ([self bugsnagStarted]) {
-        [self.client addRuntimeVersionInfo:info withKey:key];
     }
 }
 
@@ -353,15 +324,6 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
 {
     if ([self bugsnagStarted]) {
         [self.client removeOnSessionBlock:block];
-    }
-}
-
-/**
- * Intended for internal use only - sets the code bundle id for React Native
- */
-+ (void)updateCodeBundleId:(NSString *)codeBundleId {
-    if ([self bugsnagStarted]) {
-        self.client.codeBundleId = codeBundleId;
     }
 }
