+++ Bugsnag/Client/BugsnagClient+Private.h
@@ -6,9 +6,7 @@
 //  Copyright © 2020 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagClient.h>
-
-#import "BugsnagMetadata+Private.h" // For BugsnagObserverBlock
+#import "BugsnagInternals.h"
 
 @class BSGAppHangDetector;
 @class BSGEventUploader;
@@ -23,16 +21,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef NS_ENUM(NSInteger, BSGClientObserverEvent) {
-    BSGClientObserverAddFeatureFlag,    // value: BugsnagFeatureFlag
-    BSGClientObserverClearFeatureFlag,  // value: NSString
-    BSGClientObserverUpdateContext,     // value: NSString
-    BSGClientObserverUpdateMetadata,    // value: BugsnagMetadata
-    BSGClientObserverUpdateUser,        // value: BugsnagUser
-};
-
-typedef void (^ BSGClientObserver)(BSGClientObserverEvent event, _Nullable id value);
-
 @interface BugsnagClient ()
 
 #pragma mark Properties
@@ -43,22 +31,12 @@ typedef void (^ BSGClientObserver)(BSGClientObserverEvent event, _Nullable id va
 
 @property (nullable, nonatomic) BugsnagEvent *appHangEvent;
 
-@property (nullable, nonatomic) NSString *codeBundleId;
-
-@property (retain, nonatomic) BugsnagConfiguration *configuration;
-
 /// The App hang or OOM event that caused the last launch to crash.
 @property (nullable, nonatomic) BugsnagEvent *eventFromLastLaunch;
 
 @property (strong, nonatomic) BSGEventUploader *eventUploader;
 
 @property (nonatomic) NSMutableDictionary *extraRuntimeInfo;
-
-@property (strong, nonatomic) BugsnagMetadata *metadata; // Used in BugsnagReactNative
-
-@property (readonly, nonatomic) BugsnagNotifier *notifier; // Used in BugsnagReactNative
-
-@property (strong, nonatomic) BugsnagSessionTracker *sessionTracker; // Used in BugsnagReactNative
 
 @property (nonatomic) BOOL started;
 
@@ -87,27 +65,7 @@ typedef void (^ BSGClientObserver)(BSGClientObserverEvent event, _Nullable id va
 
 @property (strong, nonatomic) BugsnagSystemState *systemState;
 
-@property (readonly, atomic) BugsnagUser *user;
-
-@property (nullable, nonatomic) BSGClientObserver observer; // Used in BugsnagReactNative
-
 #pragma mark Methods
-
-- (void)addRuntimeVersionInfo:(NSString *)info withKey:(NSString *)key;
-
-- (NSDictionary *)collectAppWithState; // Used in BugsnagReactNative
-
-- (NSArray *)collectBreadcrumbs; // Used in BugsnagReactNative
-
-- (NSDictionary *)collectDeviceWithState; // Used in BugsnagReactNative
-
-- (NSArray *)collectThreads:(BOOL)unhandled; // Used in BugsnagReactNative
-
-- (BugsnagAppWithState *)generateAppWithState:(NSDictionary *)systemInfo;
-
-- (BugsnagDeviceWithState *)generateDeviceWithState:(NSDictionary *)systemInfo;
-
-- (void)notifyInternal:(BugsnagEvent *)event block:(nullable BugsnagOnErrorBlock)block;
 
 - (void)start;
