+++ Bugsnag/Configuration/BugsnagConfiguration+Private.h
@@ -6,9 +6,7 @@
 //  Copyright © 2020 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagConfiguration.h>
-
-#import "BSGFeatureFlagStore.h"
+#import "BugsnagInternals.h"
 
 @class BugsnagNotifier;
 
@@ -28,16 +26,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (copy, nonatomic) BugsnagMetadata *metadata;
 
-@property (nullable, nonatomic) BugsnagNotifier *notifier;
-
 @property (readonly, nullable, nonatomic) NSURL *notifyURL;
-
-@property (nonatomic) NSMutableArray<BugsnagOnBreadcrumbBlock> *onBreadcrumbBlocks;
-
-@property (nonatomic) NSMutableArray<BugsnagOnSendErrorBlock> *onSendBlocks;
-
-/// Hooks for modifying sessions before they are sent to Bugsnag. Intended for internal use only by React Native/Unity.
-@property (nonatomic) NSMutableArray<BugsnagOnSessionBlock> *onSessionBlocks;
 
 @property (nonatomic) NSMutableSet *plugins;
