+++ Bugsnag/Delivery/BSGEventUploader.m
@@ -16,6 +16,7 @@
 #import "BSGUtils.h"
 #import "BugsnagConfiguration.h"
 #import "BugsnagEvent+Private.h"
+#import "BugsnagInternals.h"
 #import "BugsnagLogger.h"
