+++ Bugsnag/Helpers/BSGFeatureFlagStore.h
@@ -6,11 +6,9 @@
 //  Copyright © 2021 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagFeatureFlag.h>
+#import "BugsnagInternals.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-typedef NSMutableArray<BugsnagFeatureFlag *> BSGFeatureFlagStore;
 
 void BSGFeatureFlagStoreAddFeatureFlag(BSGFeatureFlagStore *store, NSString *name, NSString *_Nullable variant);
