+++ Bugsnag/Payload/BugsnagError+Private.h
@@ -6,7 +6,7 @@
 //  Copyright © 2020 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagError.h>
+#import "BugsnagInternals.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -15,13 +15,6 @@ NS_ASSUME_NONNULL_BEGIN
 @interface BugsnagError ()
 
 - (instancetype)initWithKSCrashReport:(NSDictionary *)event stacktrace:(NSArray<BugsnagStackframe *> *)stacktrace;
-
-- (instancetype)initWithErrorClass:(NSString *)errorClass
-                      errorMessage:(nullable NSString *)errorMessage
-                         errorType:(BSGErrorType)errorType
-                        stacktrace:(nullable NSArray<BugsnagStackframe *> *)stacktrace;
-
-+ (BugsnagError *)errorFromJson:(NSDictionary *)json;
 
 /// The string representation of the BSGErrorType
 @property (copy, nonatomic) NSString *typeString;
