+++ Bugsnag/Payload/BugsnagEvent+Private.h
@@ -6,9 +6,8 @@
 //  Copyright © 2020 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagEvent.h>
-
 #import "BSGFeatureFlagStore.h"
+#import "BugsnagInternals.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -28,8 +27,6 @@ NS_ASSUME_NONNULL_BEGIN
 /// The release stages used to notify at the time this report is captured.
 @property (readwrite, copy, nullable, nonatomic) NSArray *enabledReleaseStages;
 
-@property (readwrite, strong, nonnull, nonatomic) BSGFeatureFlagStore *featureFlagStore;
-
 /// The event state (whether the error is handled/unhandled.)
 @property (readwrite, nonatomic) BugsnagHandledState *handledState;
 
@@ -48,30 +45,12 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (readwrite, nonnull, nonatomic) BugsnagUser *user;
 
-- (instancetype)initWithApp:(BugsnagAppWithState *)app
-                     device:(BugsnagDeviceWithState *)device
-               handledState:(BugsnagHandledState *)handledState
-                       user:(BugsnagUser *)user
-                   metadata:(BugsnagMetadata *)metadata
-                breadcrumbs:(NSArray<BugsnagBreadcrumb *> *)breadcrumbs
-                     errors:(NSArray<BugsnagError *> *)errors
-                    threads:(NSArray<BugsnagThread *> *)threads
-                    session:(nullable BugsnagSession *)session;
-
-- (instancetype)initWithJson:(NSDictionary *)json;
-
 - (instancetype)initWithKSReport:(NSDictionary *)KSReport;
 
 - (instancetype)initWithUserData:(NSDictionary *)event;
 
-- (void)attachCustomStacktrace:(NSArray *)frames withType:(NSString *)type; // Used in BugsnagReactNative
-
 /// Whether this report should be sent, based on release stage information cached at crash time and within the application currently.
 - (BOOL)shouldBeSent;
-
-- (void)symbolicateIfNeeded;
-
-- (NSDictionary *)toJsonWithRedactedKeys:(nullable NSSet *)redactedKeys;
 
 - (void)trimBreadcrumbs:(NSUInteger)bytesToRemove;
