+++ Bugsnag/Payload/BugsnagSession+Private.h
@@ -6,10 +6,9 @@
 //  Copyright © 2020 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagSession.h>
-
 #import "BSGDefines.h"
 #import "BSG_KSCrashReportWriter.h"
+#import "BugsnagInternals.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -27,11 +26,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 #pragma mark Properties
 
-@property (nonatomic) NSUInteger handledCount;
-
 @property (getter=isStopped, nonatomic) BOOL stopped;
-
-@property (nonatomic) NSUInteger unhandledCount;
 
 @property (readwrite, nonnull, nonatomic) BugsnagUser *user;
