+++ Bugsnag/Payload/BugsnagThread+Private.h
@@ -6,8 +6,8 @@
 //  Copyright © 2020 Bugsnag Inc. All rights reserved.
 //
 
-#import <Bugsnag/BugsnagThread.h>
 #import "BSGDefines.h"
+#import "BugsnagInternals.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -22,10 +22,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (instancetype)initWithThread:(NSDictionary *)thread binaryImages:(NSArray *)binaryImages;
 
-+ (NSArray<BugsnagThread *> *)allThreads:(BOOL)allThreads callStackReturnAddresses:(NSArray<NSNumber *> *)callStackReturnAddresses;
-
-+ (instancetype)threadFromJson:(NSDictionary *)json;
-
 @property (readonly, nullable, nonatomic) NSString *crashInfoMessage;
 
 @property (readwrite, nonatomic) BOOL errorReportingThread;
@@ -35,8 +31,6 @@ NS_ASSUME_NONNULL_BEGIN
 #if BSG_HAVE_MACH_THREADS
 + (nullable instancetype)mainThread;
 #endif
-
-+ (NSMutableArray *)serializeThreads:(nullable NSArray<BugsnagThread *> *)threads;
 
 + (NSMutableArray<BugsnagThread *> *)threadsFromArray:(NSArray *)threads binaryImages:(NSArray *)binaryImages;
